backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==2.25.0
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.3.0
urllib3==2.2.1
pydantic==2.7.0
